fix: change router view page file's location
