+++ src/components/Pages/Home/HomeSkillsBrief.tsx
@@ -2,11 +2,10 @@ import React from 'react'
 import clsx from 'clsx'
 import Toolbar from '@material-ui/core/Toolbar'
 import { AnimationOnScroll } from 'react-animation-on-scroll'
-import { Title, Text } from '../../components/Common'
-import { useWindowSize } from '../../hooks'
+import { Title, Text } from '../../Common'
+import { useWindowSize } from '../../../hooks'
 
-// 该页面组件在屏幕中居中展示
-const HomeWorksBrief = () => {
+const HomeSkillsBrief = () => {
   const { isMobile } = useWindowSize()
 
   return (
@@ -17,31 +16,31 @@ const HomeWorksBrief = () => {
       <Toolbar variant="dense" />
 
       {/* title */}
-      <AnimationOnScroll animateIn="animate__fadeInRight" animateOnce={true}>
-        <Title title={'作品集について'} />
+      <AnimationOnScroll animateIn="animate__fadeInLeft" animateOnce={true}>
+        <Title title={'スキルについて'} />
       </AnimationOnScroll>
 
       {/* content */}
       <Text align={isMobile ? 'left' : 'center'}>
         <AnimationOnScroll
-          animateIn="animate__fadeInLeft"
+          animateIn="animate__fadeInRight"
           animateOnce={true}
           delay={100}
         >
           <p className={clsx(isMobile ? 'display-inline' : 'display-block')}>
-            フロントからバックエンドまで幅広く練習してきました。
+            フロントからバックエンドまで幅広くスキルアップに取り組んできました。
           </p>
           <p className={clsx(isMobile ? 'display-inline' : 'display-none')}>
-            これからポロジェットに携わりながら、作品集を不定期に更新します。
+            広く浅くではありますが、今後も幅広い領域に挑戦して引き続き頑張りたいと考えています
           </p>
         </AnimationOnScroll>
         <AnimationOnScroll
-          animateIn="animate__fadeInLeft"
+          animateIn="animate__fadeInRight"
           animateOnce={true}
           delay={200}
         >
           <p className={clsx(isMobile ? 'display-none' : 'display-block')}>
-            これからポロジェットに携わりながら、作品集を不定期に更新します。
+            広く浅くではありますが、今後も幅広い領域に挑戦して引き続き頑張りたいと考えています
           </p>
         </AnimationOnScroll>
       </Text>
@@ -49,4 +48,4 @@ const HomeWorksBrief = () => {
   )
 }
 
-export default HomeWorksBrief
+export default HomeSkillsBrief
