+++ src/pages/about.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import logo from '../../assets/logo/favicon.ico'
+import logo from '../assets/logo/favicon.ico'
 
 const IndexPage = (): JSX.Element => {
   return (
